// src/context/photos/hooks/use-photo.ts
import { useQuery, useQueryClient } from "@tanstack/react-query";
import { toast } from "sonner";
import { fetcher } from "../../../helpers/api";
import type { Photo } from "../models/photo";
import type { PhotoNewFormSchema } from "../schemas";
import { api } from './../../../helpers/api';

interface PhotoDetailsResponse extends Photo {
  nextPhotoId?: string;
  previousPhotoId?: string;
}

export default function usePhoto(id?: string){
  const {data, isLoading } = useQuery<PhotoDetailsResponse>({
    queryKey: ["photos", id],
    queryFn: ()=> fetcher(`/photos/${id}`),
    enabled: !!id
  });

  const queryClient = useQueryClient();

  async function createPhoto(payload: PhotoNewFormSchema) {
    try {
      const {data: photo} = await api.post<Photo>("/photos", {title: payload.title });

      await api.post(`/photos/${photo?.id}/image`, { 
        file: payload.file[0]}, {
          headers: {
            "Content-Type": "multipart/form-data"
          }
        });

        if(payload.albumsIds && payload.albumsIds.length > 0){
          await api.put(`/photos/${photo?.id}/albums`, {
            albumsIds: payload.albumsIds
          });
        }

      queryClient.invalidateQueries({queryKey: ["photos"]});

      toast.success("Foto criada com sucesso");
    } catch (error) {
      toast.error("Erro ao criar foto");
      throw error;
    }
  }

  return {
    photo: data,
    nextPhotoId: data?.nextPhotoId,
    previousPhotoId: data?.previousPhotoId,
    isLoadingPhoto: isLoading,
    createPhoto
  }
}